fix (most) test issues after renaming
