+++ openpathsampling/experimental/simstore/__init__.py
@@ -5,3 +5,8 @@ from .storage import (
 from .class_info import ClassInfoContainer, ClassInfo
 from .sql_backend import SQLStorageBackend
 from . import dict_serialization_helpers
+from .storable_functions import (
+    Processor, StorableFunctionConfig, StorableFunctionResults,
+    StorableFunction, StorageFunctionHandler, storable_function_find_uuids
+)
+from .callable_codec import CallableCodec
