+++ openpathsampling/experimental/simstore/callable_codec.py
