+++ openpathsampling/experimental/simstore/memory_backend.py
