+++ openpathsampling/experimental/simstore/test_callable_codec.py
