+++ openpathsampling/experimental/storage/test_collective_variables.py
@@ -11,9 +11,9 @@ from openpathsampling.tests.test_helpers import make_1d_traj, data_filename
 from openpathsampling.engines import openmm as ops_omm
 
 from .collective_variables import *
-from .test_storable_function import MockBackend
-from .storable_functions import StorageFunctionHandler
-from .serialization_helpers import get_uuid
+from ..simstore.test_storable_function import MockBackend
+from ..simstore import StorageFunctionHandler
+from ..simstore.serialization_helpers import get_uuid
 
 
 try:
